Fix issue where data was incorrectly labeled as linear
This issue was caused by arrays with small numbers that wasn't
caught by Numpy's allclose function which has a default tolerance
of 1e-8. The tolerance is not set to 1e-15 but this isn't a
perfect solution.
Also, the function get_extent_of_data_dim now uses nanmin, nanmax
for all arrays so that it does not depend on whether the array is
determined to be linear.